# docs/openapi_spec.yaml
openapi: 3.0.0
info:
  description: APIs for Vending Machine Service
  version: "1.0.0"
  title: Vending Machine Service
  contact:
    email: vengateshs@gmail.com
servers:
  - url: 'http://127.0.0.1'
...
paths:
...
  /$deposit:
    post:
      tags:
        - Actions
      summary: Deposit coins into user's vending machine account
      description: Users with a “buyer” role can deposit 5, 10, 20, 50 and 100 cent coins into their vending machine account
      operationId: depositCoins
      security:
        - basicAuth: []      
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
        required: true
      responses:
        "200":
          description: Coins deposited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Entity Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "406":
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "415":
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "503":
          description: Service Unavilable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
  /$buy:
    post:
      tags:
        - Actions
      summary: Buy products from the money deposited
      description: Accepts productId and amount of products so users with a “buyer” role can buy products with the money they’ve deposited. API should return total they’ve spent, products they’ve purchased and their change if there’s any (in 5, 10, 20, 50 and 100 cent coins)
      operationId: buyProducts
      security:
        - basicAuth: []      
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyProductsRequest'
        required: true
      responses:
        "200":
          description: Products bought successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyProductsResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Entity Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "406":
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "415":
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "503":
          description: Service Unavilable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
  /$reset:
    post:
      tags:
        - Actions
      summary: Reset deposit
      description: Users with a “buyer” role can reset their deposit
      operationId: resetDeposit
      security:
        - basicAuth: []      
      responses:
        "200":
          description: Deposit got reset successfully
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Entity Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "503":
          description: Service Unavilable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: double
        createdBy:
          type: string
...
    ErrorResponse:
      properties:
        error:
          description: A user friendly error message
          example: An example error message.
          type: string
        errorCode:
          description: A machine-friendly identifier for the error response
          type: string
        traceId:
          description: A unique identifier for the request
          type: string
      required:
      - error
      - errorCode
      - traceId
    CoinValueEnum:
      type: integer
      enum:
      - 5
      - 10
      - 20
      - 50
      - 100
    DepositRequest:
      type: object
      properties:
        coinValue:
          $ref: "#/components/schemas/CoinValueEnum"         
    DepositResponse:
      type: object
      properties:
        balance:
          items:
            $ref: '#/components/schemas/BalanceCoins'
          type: array         
    BuyProductsRequest:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        amount:
          type: integer
          format: int64
    BalanceCoins:
      type: object
      properties:
        coinValue:
          $ref: "#/components/schemas/CoinValueEnum"
        amount:
          type: integer
          format: int64
    BuyProductsResponse:
      type: object
      properties:
        totalSpent:
          type: number
          format: double
        product:
          $ref: '#/components/schemas/Product'
        amount:
          type: integer
          format: int64
        change:
          items:
            $ref: '#/components/schemas/BalanceCoins'
          type: array         
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic